Update mips code to allow building testing tool
  Still needs significant work, but very basic tests works now (only
hello.tst and bottles.tst tested). But still needs to write logic to
properly flush cache, otherwise, if not "disassembling" the code
before execution, if may SIGTRAP or SIGILL.

+++ lightning.h
@@ -170,7 +170,7 @@ struct jit_local_state {
     int		 alloca_offset;
     int		 stack_length;
     int		 stack_offset;
-    short	*stack;
+    int		*stack;
 };
 #else
 #  error GNU lightning does not support the current target
